implement requested changes
